feat: day 10, part 2 done, phewwww
Thank god I found this video:
https://www.youtube.com/watch?v=p-xa-3V5KO8&amp=&feature=youtu.be

Otherwise I could not have understand the question
as I was too focused on grid mechanics. The last time
I tried math instead of grod mechanics (previous day 3, wires)
it was wrong so I didn't want to think of Math this time. What a mistake.
Math.atan2 was my best friend in my earliest Flash days tinkering around
with objects following the mouse.

In the end it worked out and this was a pretty nice challenge.